# openapi/v1.yaml
openapi: 3.0.3
info:
  title: Procezo API
  version: 1.4.0
paths:
...
  /api/v1/decisions/{id}/:
    get:
      operationId: decisions_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - decisions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
          description: ''
  /api/v1/decisions/{id}/historique/:
    get:
      operationId: decisions_historique_list
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - decisions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDecisionEventList'
          description: ''
  /api/v1/decisions/{id}/retourner/:
    post:
      operationId: decisions_retourner_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - decisions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionReturn'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
          description: ''
  /api/v1/decisions/{id}/valider/:
    post:
      operationId: decisions_valider_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - decisions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionTransition'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
          description: ''
...
  /api/v1/demandes/{id}/:
    get:
      operationId: demandes_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - demandes
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
          description: ''
    patch:
      operationId: demandes_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - demandes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
          description: ''
...
  /api/v1/dossiers/{id}/:
    get:
      operationId: dossiers_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - dossiers
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
          description: ''
    patch:
      operationId: dossiers_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - dossiers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseUpdate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
          description: ''
...
components:
  schemas:
    Act:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        request_version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        mode:
          $ref: '#/components/schemas/ModeEnum'
        imported_document:
          type: string
          format: uuid
          nullable: true
        sha256:
          type: string
          maxLength: 64
        prepared_by:
          type: integer
        prepared_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - id
      - mode
      - prepared_at
      - prepared_by
      - request_version
      - sha256
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        reference:
          type: string
          readOnly: true
        unit:
          type: string
          format: uuid
          readOnly: true
        unit_code:
          type: string
          readOnly: true
        classification:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        assignee:
          type: integer
          readOnly: true
        assignee_username:
          type: string
          readOnly: true
        next_action:
          type: string
          readOnly: true
        version:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - assignee
      - assignee_username
      - classification
      - created_at
      - id
      - next_action
      - reference
      - status
      - unit
      - unit_code
      - updated_at
      - version
...
    CaseUpdate:
      type: object
      properties:
        version:
          type: integer
          minimum: 1
        next_action:
          type: string
          maxLength: 240
        status:
          $ref: '#/components/schemas/StatusEnum'
      required:
      - version
...
    Decision:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        case:
          type: string
          format: uuid
        author:
          type: integer
        kind:
          $ref: '#/components/schemas/KindEnum'
        reason:
          type: string
          maxLength: 1000
        request_assessment:
          type: string
          format: uuid
          nullable: true
        observation_assessment:
          type: string
          format: uuid
          nullable: true
        replaces:
          type: string
          format: uuid
          nullable: true
        state:
          $ref: '#/components/schemas/DecisionStateEnum'
        version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        validator:
          type: integer
          nullable: true
        return_comment:
          type: string
          maxLength: 500
        validated_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        prototype_only:
          type: boolean
      required:
      - author
      - case
      - created_at
      - id
      - kind
      - reason
      - updated_at
...
    DecisionEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        kind:
          type: string
          maxLength: 20
        actor:
          type: integer
        comment:
          type: string
          maxLength: 500
        version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - actor
      - created_at
      - id
      - kind
      - version
    DecisionReturn:
      type: object
      properties:
        version:
          type: integer
          minimum: 1
        comment:
          type: string
          maxLength: 500
      required:
      - comment
      - version
    DecisionStateEnum:
      enum:
      - proposed
      - returned
      - validated
      type: string
      description: |-
        * `proposed` - Proposée
        * `returned` - Retournée
        * `validated` - Validée
    DecisionTransition:
      type: object
      properties:
        version:
          type: integer
          minimum: 1
      required:
      - version
...
    Issuance:
      type: object
      properties:
        act:
          type: string
          format: uuid
        signature_proof:
          type: string
          format: uuid
        dispatch_proof:
          type: string
          format: uuid
        issued_at:
          type: string
          format: date-time
        recorded_at:
          type: string
          format: date-time
          readOnly: true
        prototype_only:
          type: boolean
      required:
      - act
      - dispatch_proof
      - issued_at
      - recorded_at
      - signature_proof
...
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        number:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        label:
          type: string
          maxLength: 500
        assessment_version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
      required:
      - id
      - label
      - number
    ItemInput:
      type: object
      properties:
        label:
          type: string
          maxLength: 500
      required:
      - label
    KindEnum:
      enum:
      - classification
      - mission
      - complement
      - gelec
      type: string
      description: |-
        * `classification` - Classement sans suite
        * `mission` - Mission complémentaire
        * `complement` - Demande de complément
        * `gelec` - Relais GELEC
...
    ModeEnum:
      enum:
      - generated
      - imported
      type: string
      description: |-
        * `generated` - Format Procezo
        * `imported` - Document de l'agent
...
    PaginatedDecisionEventList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/DecisionEvent'
...
    Request:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        case:
          type: string
          format: uuid
        author:
          type: integer
        target_type:
          $ref: '#/components/schemas/TargetTypeEnum'
        target_name:
          type: string
          maxLength: 240
        represented_name:
          type: string
          maxLength: 240
        subject:
          type: string
          maxLength: 240
        mode:
          $ref: '#/components/schemas/ModeEnum'
        due_on:
          type: string
          format: date
          nullable: true
        state:
          $ref: '#/components/schemas/RequestStateEnum'
        version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
          readOnly: true
        acts:
          type: array
          items:
            $ref: '#/components/schemas/Act'
          readOnly: true
        issuance:
          allOf:
          - $ref: '#/components/schemas/Issuance'
          readOnly: true
      required:
      - acts
      - author
      - case
      - created_at
      - id
      - issuance
      - items
      - mode
      - subject
      - target_name
      - target_type
      - updated_at
...
    RequestStateEnum:
      enum:
      - draft
      - submitted
      - validated
      - signed
      - issued
      type: string
      description: |-
        * `draft` - Brouillon
        * `submitted` - Soumis
        * `validated` - Validé
        * `signed` - Signature constatée
        * `issued` - Émis
    RequestUpdate:
      type: object
      properties:
        version:
          type: integer
          minimum: 1
        target_type:
          $ref: '#/components/schemas/TargetTypeEnum'
        target_name:
          type: string
          maxLength: 240
        represented_name:
          type: string
          maxLength: 240
        subject:
          type: string
          maxLength: 240
        mode:
          $ref: '#/components/schemas/ModeEnum'
        due_on:
          type: string
          format: date
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemInput'
      required:
      - version
...
    StatusEnum:
      enum:
      - open
      - in_progress
      type: string
      description: |-
        * `open` - Ouvert
        * `in_progress` - En cours
...
    TargetTypeEnum:
      enum:
      - broker
      - organization
      type: string
      description: |-
        * `broker` - Commissionnaire en douane
        * `organization` - Entreprise ou ONG
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid